requirements/base.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==3.7.1
asgiref==3.10.0
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
billiard==4.2.1
black==25.9.0
celery==5.4.0
certifi==2025.10.5
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
constantly==23.10.4
coverage==7.10.7
cron_descriptor==2.0.6
cryptography==44.0.2
daphne==4.1.2
Django==5.1.1
django-celery-beat==2.7.0
django-cors-headers==4.4.0
django-timezone-field==7.1
django-redis==6.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
iniconfig==2.1.0
jsonpatch==1.33
jsonpointer==3.0.0
kombu==5.5.0
langchain==1.2.0
langchain-core==1.2.1
langchain-huggingface==1.2.0
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.0
langsmith==0.5.0
msgpack==1.1.2
mypy_extensions==1.1.0
ollama==0.6.1
orjson==3.11.5
ormsgpack==1.12.1
packaging==24.2
pathspec==0.12.1
pika==1.3.2
platformdirs==4.5.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2==2.9.10
peft==0.18.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pyOpenSSL==25.0.0
pytest==8.4.2
pytest-cov==7.0.0
pytest-django==4.11.1
pytest-mock==3.15.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
pytokens==0.1.10
PyYAML==6.0.3
redis==5.2.1
requests==2.32.5
requests-toolbelt==1.0.0
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.1
tenacity==9.1.2
Twisted==24.11.0
txaio==23.1.1
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.9.1
transformers==4.57.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.1
urllib3==2.5.0
uuid_utils==0.12.0
vine==5.1.0
watchgod==0.8.2
wcwidth==0.2.14
xxhash==3.6.0
zope.interface==8.0.1
zstandard==0.25.0
